requirements.txt
fastapi
uvicorn
openai
supabase
python-dotenv
faster-whispe
python-multipart